+++ .gitignore
@@ -1,5 +1,31 @@
-out
-dist
-node_modules
-.vscode-test/
+# Dependencies
+/node_modules
+
+# Build output
+/out
+/dist
 *.vsix
+
+# VS Code extension testing
+.vscode-test/
+
+# Vite cache
+.vite/
+
+# Log files
+*.log
+npm-debug.log*
+yarn-error.log
+pnpm-debug.log
+
+# OS generated files
+.DS_Store
+Thumbs.db
+
+# Editor directories and files
+.idea/
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.swp
