+++ media/main.js
@@ -1,8 +1,11 @@
 import { createApp } from 'vue';
+import { createPinia } from 'pinia';
 import App from './App.vue';
 
 // @ts-ignore
 const vscode = acquireVsCodeApi();
+const pinia = createPinia();
 
 const app = createApp(App, { vscode });
+app.use(pinia);
 app.mount('#app');
