+++ package.json
@@ -59,5 +59,8 @@
     "typescript": "^5.4.5",
     "vite": "^5.4.19",
     "vue": "^3.4.27"
+  },
+  "dependencies": {
+    "pinia": "^3.0.3"
   }
 }
